Switch from actix to axum.
Reduces unstripped release binary size from 5924536 to 3617328
and resident memory use from 5452 to 3612 KB.

Tiny increase in lines of code mostly due to formatting of
the error handling on get_service. nbd.

(This change is desirable for my application because
this binary runs on a small embedded system that is a
gateway for my shark power monitor. Your mileage may vary.)